Simplify playerargs setting
prefer-ipv4, nolirc and really-quiet flags are added automatically when needed
they no longer appear in playerargs and are handled out of sight of the user

Similarly, the "set fullscreen" option no longer affects the playerargs setting
and is passed to the player behind-the-scenes if set.

Custom playerargs can still be set by the user for controlling any other
aspects of the player such as window size, geometry etc.

This makes coding the config section easier as when the user sets the player or
the fullscreen options, no other changes need to be made.