Methods to utilize search result count info for pagination.
Added methods for extracting search result count and pagination
data from json, and for formatting current page information.

With the total amount of search results present, it can be determined
how many pages of search results can be made available to the user
(considering the current terminal size).

Note: New API limits the number of search result items it provides to the
first 500.