Implements playlist generation based on video descriptions
Extends mkp functionality by having it act on the --description parameter.

When the --description parameter is passed, the software looks through g.model for suitable datatypes. It will find the first applicable object based on numerical input from the user, download that videos description and parse it for artist - track combinations, and will store that in a local playlist using the existing generate_playlist functionality.