Revert "Allow more than 500 results for searches"
This reverts commit e378dbcfd99f6188b82679769bcc662c5e0be879.

According to the docs:

"Note: Search results are constrained to a maximum of 500 videos if your
request specifies a value for the channelId parameter and sets the type
parameter value to video, but it does not also set one of the
forContentOwner, forDeveloper, or forMine filters."

https://developers.google.com/youtube/v3/docs/search/list

The number of results returned does, however, seem limited for a normal
search as well, and not entirely consistently. If possible, this should
be fixed, but I don't know what is going on.