Fix cases of config values not being checked properly (#1326)
* Fix delay on play if subtitles not enabled

It seems like the intent here was to check that subtitles were enabled,
not that the config items exist. This fixes a needless delay to fetch
the subtitles on playback if subtitles are not enabled.

* Fix incorrect error if user not found

Fixing a mistake that causes this error to always be displayed even when
the search_music config option isn't enabled.